mixmaster_ai_backend/requirements.txt
# Core dependencies
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.4.0

# Audio processing
librosa>=0.9.0
pydub>=0.25.1
soundfile>=0.10.3
audioread>=2.1.9
essentia>=2.1.0

# Audio mastering
matchering>=2.0.6

# ML models (optional)
torch>=1.9.0
torchaudio>=0.9.0

# API and web server
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
aiofiles>=0.7.0

# Utilities
tqdm>=4.62.0
ffmpeg-python>=0.2.0
